+++ guest/firmware/paging.h
@@ -1,7 +1,7 @@
 // paging.h
 #pragma once
 
-#include <cstdint>
+#include <stdint.h>
 #define PAGE_PRESENT  (1 << 0)
 #define PAGE_WRITE    (1 << 1)
 #define PAGE_HUGE     (1 << 7)  // 2MB pages in PD entries
@@ -20,6 +20,6 @@ static inline void paging_init(void) {
 
     // PD[0..511] → identity map first 1GB in 2MB chunks
     for (int i = 0; i < 512; i++) {
-        pd[i] = ((uint64_t)i * 0x200000) | PAGE_PRESENT | PAGE_WRITE | PAGE_HUGE;
+        pd[i] = ((uint32_t)i * 0x200000) | PAGE_PRESENT | PAGE_WRITE | PAGE_HUGE;
     }
 }
